runtime cpuflag for Intel MULX, cleaned in tfm.c
